doks/requirements.txt
boto3==1.37.20
fire==0.7.0
python-dotenv==1.1.0
tqdm==4.67.1